dags/requirements.txt
psycopg2
pandas
boto3
apache-airflow-providers-amazon